Press and hold to capture expression functionality.
